feat(api): replace hand-rolled Fastify API with @wolfstar/plugin-api
Fold the guild-config REST API into the same @wolfstar/http-framework
piece-loading system already used for commands, dropping the one-off
Fastify integration in favor of the official plugin. Routes move from
src/api/routes/ (manual container.server.route() calls, Fastify
request/reply) to src/routes/ (auto-loaded Route pieces, file-based
routing), with a shared requireMapping() helper replacing the
duplicated inline bearer-token checks.

Requires bumping @wolfstar/http-framework to ^3.0.0 (plugin-api's peer
dependency) and @wolfstar/shared-http-pieces to ^1.2.6 to match, plus a
pnpm override for @wolfstar/http-framework-i18n's broken workspace:
dependency on @wolfstar/i18next-backend.
