+++ .github/copilot-instructions.md
@@ -4,8 +4,8 @@
 - Ring is a private centralized management android for the WolfStar Network,
   built with TypeScript. It uses HTTP interactions via Discord's HTTP-based bot
   architecture (`@wolfstar/http-framework`) rather than a persistent WebSocket
-  connection. It exposes internal REST API endpoints via Fastify for other bots
-  in the network to query guild configuration.
+  connection. It exposes internal REST API endpoints via `@wolfstar/plugin-api`
+  for other bots in the network to query guild configuration.
 - Always reference these instructions first and fall back to search or
   documentation queries only when you encounter unexpected information.
 
@@ -32,7 +32,7 @@
 
 | Type             | Convention      | Example                             |
 | ---------------- | --------------- | ----------------------------------- |
-| Directories      | kebab-case      | `api/routes/`                       |
+| Directories      | kebab-case      | `routes/`                           |
 | TypeScript files | camelCase       | `config.ts`                         |
 | Variables        | camelCase       | `guildId`, `mappings`               |
 | Constants        | PascalCase enum | `Mappings.staryl`                   |
@@ -44,7 +44,7 @@
 
 ## Import Conventions
 
-- Use TypeScript path mapping aliases for internal imports: `#lib/*`, `#api/*`,
+- Use TypeScript path mapping aliases for internal imports: `#lib/*`,
   `#common/*`, `#types`
 - Use `type` imports for type-only values: `import type { ... } from "..."`
 - Group imports: type imports first, then internal aliases, then external
@@ -55,11 +55,14 @@
 
 ### Key Patterns
 
-- **HTTP Framework**: Built on `@wolfstar/http-framework` (Fastify-based),
-  handling Discord interactions via HTTP endpoints instead of WebSocket gateway
-- **Internal API**: Fastify REST endpoints (`src/api/routes/`) that serve guild
-  configuration data to other bots in the WolfStar Network (Staryl, WolfStar),
-  authenticated via bearer tokens
+- **HTTP Framework**: Built on `@wolfstar/http-framework`, handling Discord
+  interactions via HTTP endpoints instead of WebSocket gateway
+- **Internal API**: `@wolfstar/plugin-api` REST endpoints (`src/routes/`) that
+  serve guild configuration data to other bots in the WolfStar Network (Staryl,
+  WolfStar), authenticated via bearer tokens. Routes are `Route` pieces
+  auto-loaded from `src/routes/`, with the path/method inferred from file
+  location (folder segments, `[param]` dynamic segments, and a `.<method>`
+  filename suffix)
 - **Guild limits**: Canonical defaults, ranges, and bot mappings live in
   `src/lib/common/limits.ts` (`LimitDefinitions`, `Mappings`, `getMappings`)
 - **Database**: PostgreSQL with Prisma ORM. Models use `@@map()` for snake_case
@@ -67,16 +70,17 @@
   must stay in sync with `LimitDefinitions`
 - **i18n**: Multi-language support via `@wolfstar/http-framework-i18n` with
   locale JSON files in `src/locales/`
-- **Container Pattern**: Services (Prisma, Fastify, Logger) are attached to
-  `container` from `@sapphire/pieces` with corresponding type augmentations
+- **Container Pattern**: Services (Prisma, the `@wolfstar/plugin-api` API
+  server, Logger) are attached to `container` from `@sapphire/pieces` with
+  corresponding type augmentations
 
 ### Directory Structure
 
 - `src/main.ts` - Application entry point
-- `src/api/routes/` - Internal REST API endpoints (guild config queries)
-- `src/api/routes/_load.ts` - Route loader (imports all route files)
+- `src/routes/` - Internal REST API endpoints (guild config queries),
+  auto-loaded as `@wolfstar/plugin-api` `Route` pieces
 - `src/commands/` - Discord slash commands using decorator pattern
-- `src/lib/setup/` - Application initialization (env, Fastify, Prisma, logger)
+- `src/lib/setup/` - Application initialization (env, Prisma, logger)
 - `src/lib/common/` - Shared constants and guild limit definitions
 - `src/lib/types/` - TypeScript type definitions and env augmentations
 - `src/locales/` - Translation JSON files organized by locale
@@ -106,17 +110,44 @@ export class UserCommand extends Command {
 
 ### API Route Structure
 
-Routes are registered directly on `container.server` (Fastify):
+Routes are `Route` pieces loaded from `src/routes/`. The path and method are
+inferred from the file's location: folder segments become path segments,
+`[param]` folders become dynamic segments, `index` collapses into its parent,
+and a `.<method>` filename suffix selects the HTTP method (e.g.
+`src/routes/guilds/[id].get.ts` registers `GET /guilds/[id]`):
 
 ```typescript
-container.server.route({
-	url: "/path",
-	method: "GET",
-	handler: async (request, reply) => {
-		// Handle request
-	},
-});
+import { requireMapping } from "#lib/api/auth";
+import { HttpCodes } from "@wolfstar/http-framework";
+import { Route, type ApiRequest, type ApiResponse } from "@wolfstar/plugin-api";
+
+export class ExampleRoute extends Route {
+	public constructor(context: Route.LoaderContext) {
+		super(context, { name: "example-get" });
+	}
+
+	public async run(request: ApiRequest, response: ApiResponse) {
+		const mappings = requireMapping(request, response);
+		if (mappings === null) return;
+
+		response.json({ success: true }, HttpCodes.OK);
+	}
+}
 ```
+
+Two non-obvious gotchas:
+
+- **Status codes go through `response.json(body, statusCode)`, not
+  `response.status(code).json(body)`.** `ApiResponse#json()` defaults its
+  `statusCode` parameter to `HttpCodes.OK`, so chaining `.status(x).json(y)`
+  silently overwrites `x` back to 200 — this fails with no error, only a wrong
+  status code on the wire.
+- **Every route needs an explicit unique `name`.** `@sapphire/pieces` keys its
+  route store by filename alone (not full path), so two files named e.g.
+  `index.get.ts` in different directories collide — the later-loaded one
+  silently wins and the other's route never gets registered (no error, no
+  warning, the endpoint just 405s). Pass `{ name: "<unique-slug>" }` in the
+  constructor's `super()` call for every route file, as shown above.
 
 ## Development Commands
 
@@ -148,7 +179,8 @@ Types: `feat`, `fix`, `refactor`, `test`, `chore`, `docs`, `style`, `perf`,
 ## Key Dependencies
 
 - `@wolfstar/http-framework` - Discord HTTP interaction framework
-  (Fastify-based)
+- `@wolfstar/plugin-api` - Standalone REST API server plugin (`Route`/
+  `Middleware` pieces) for `@wolfstar/http-framework`
 - `@wolfstar/http-framework-i18n` - Internationalization for the HTTP framework
 - `@wolfstar/shared-http-pieces` - Shared command registration and Sentry
   integration
@@ -157,7 +189,6 @@ Types: `feat`, `fix`, `refactor`, `test`, `chore`, `docs`, `style`, `perf`,
   `isNullishOrEmpty`)
 - `@discordjs/builders` - Discord slash command option builders
 - `@prisma/client` - Database ORM
-- `fastify` - HTTP server (underlying `@wolfstar/http-framework`)
 
 ## Troubleshooting
 
@@ -171,4 +202,4 @@ Types: `feat`, `fix`, `refactor`, `test`, `chore`, `docs`, `style`, `perf`,
   `RingLimits` / `DefaultLimits` in Staryl's `src/lib/utilities/ring.ts`
 
 **When in doubt:** Copy existing patterns from similar files (e.g.,
-`src/commands/`, `src/api/routes/`) before inventing new ones.
+`src/commands/`, `src/routes/`) before inventing new ones.
