+++ package.json
@@ -8,7 +8,6 @@
 	"type": "module",
 	"main": "dist/main.mjs",
 	"imports": {
-		"#api/*": "./dist/api/*.mjs",
 		"#common/*": "./dist/lib/common/*.mjs",
 		"#generated/prisma": "./dist/generated/prisma/client.mjs",
 		"#lib/*": "./dist/lib/*.mjs",
@@ -36,13 +35,13 @@
 		"@sapphire/result": "^2.8.0",
 		"@sapphire/utilities": "^3.18.2",
 		"@wolfstar/env-utilities": "^2.0.2",
-		"@wolfstar/http-framework": "^2.2.2",
+		"@wolfstar/http-framework": "^3.0.0",
 		"@wolfstar/http-framework-i18n": "^1.2.2",
 		"@wolfstar/logger": "^2.0.5",
-		"@wolfstar/shared-http-pieces": "^1.2.4",
+		"@wolfstar/plugin-api": "^1.0.0",
+		"@wolfstar/shared-http-pieces": "^1.2.6",
 		"@wolfstar/start-banner": "^2.0.3",
 		"discord-api-types": "^0.38.47",
-		"fastify": "^5.8.5",
 		"gradient-string": "^3.0.0",
 		"husky": "^9.1.7",
 		"ioredis": "^5.11.1",
