+++ pnpm-workspace.yaml
@@ -1,6 +1,7 @@
 overrides:
   ansi-regex: ^6.2.2
   minimist: ^1.2.8
+  "@wolfstar/http-framework-i18n>@wolfstar/i18next-backend": ^2.0.8
 
 allowBuilds:
   "@prisma/client": true
