--- src/api/routes/_load.ts
@@ -1,6 +1,0 @@
-import "#api/routes/index";
-import "#api/routes/guilds/[...id]";
-import "#api/routes/experiments/index";
-import "#api/routes/experiments/[...key]";
-import "#api/routes/experiments/[key]";
-import "#api/routes/experiments/overrides";
