--- src/api/routes/index.ts
@@ -1,7 +1,0 @@
-import { container } from "@sapphire/pieces";
-
-container.server.route({
-	url: "/",
-	method: "GET",
-	handler: () => ({ data: "Hello world" }),
-});
