--- src/lib/setup/fastify.ts
@@ -1,12 +1,0 @@
-import { container } from "@wolfstar/http-framework";
-import fastify from "fastify";
-
-const server = fastify();
-
-container.server = server;
-
-declare module "@sapphire/pieces" {
-	interface Container {
-		server: typeof server;
-	}
-}
